Refactor code to improve encapsulation
