+++ src/org/jacorb/orb/Delegate.java
@@ -41,9 +41,12 @@ import org.jacorb.orb.giop.RequestOutputStream;
 import org.jacorb.orb.iiop.IIOPProfile;
 import org.jacorb.orb.miop.MIOPProfile;
 import org.jacorb.orb.policies.PolicyManager;
+import org.jacorb.orb.portableInterceptor.ClientInterceptorHandler;
 import org.jacorb.orb.portableInterceptor.ClientInterceptorIterator;
 import org.jacorb.orb.portableInterceptor.ClientRequestInfoImpl;
+import org.jacorb.orb.portableInterceptor.DefaultClientInterceptorHandler;
 import org.jacorb.orb.portableInterceptor.InterceptorManager;
+import org.jacorb.orb.portableInterceptor.NullClientInterceptorHandler;
 import org.jacorb.orb.portableInterceptor.ServerInterceptorIterator;
 import org.jacorb.orb.portableInterceptor.ServerRequestInfoImpl;
 import org.jacorb.orb.util.CorbaLoc;
