+++ src/org/jacorb/orb/portableInterceptor/DefaultClientInterceptorHandler.java
@@ -19,14 +19,15 @@
  *
  */
 
-package org.jacorb.orb;
+package org.jacorb.orb.portableInterceptor;
 
 import java.util.Collection;
 import java.util.Map;
+
+import org.jacorb.orb.ApplicationExceptionHelper;
+import org.jacorb.orb.Delegate;
+import org.jacorb.orb.SystemExceptionHelper;
 import org.jacorb.orb.giop.ReplyInputStream;
-import org.jacorb.orb.portableInterceptor.ClientInterceptorIterator;
-import org.jacorb.orb.portableInterceptor.ClientRequestInfoImpl;
-import org.jacorb.orb.portableInterceptor.InterceptorManager;
 import org.omg.CORBA.portable.ApplicationException;
 import org.omg.CORBA.portable.RemarshalException;
 import org.omg.GIOP.ReplyHeader_1_2;
